+++ crates/store/src/state.rs
@@ -22,7 +22,7 @@ use miden_objects::{
     AccountError,
     account::{AccountDelta, AccountHeader, AccountId, StorageSlot},
     block::{
-        AccountTree, AccountWitness, BlockChain, BlockHeader, BlockInputs, BlockNumber,
+        AccountTree, AccountWitness, BlockHeader, BlockInputs, BlockNumber, Blockchain,
         NullifierTree, NullifierWitness, ProvenBlock,
     },
     crypto::{
@@ -30,7 +30,7 @@ use miden_objects::{
         merkle::{Mmr, MmrDelta, MmrProof, PartialMmr, SmtProof},
     },
     note::{NoteDetails, NoteId, Nullifier},
-    transaction::{OutputNote, PartialBlockChain},
+    transaction::{OutputNote, PartialBlockchain},
     utils::Serializable,
 };
 use tokio::{
@@ -62,7 +62,7 @@ pub struct TransactionInputs {
 /// Container for state that needs to be updated atomically.
 struct InnerState {
     nullifier_tree: NullifierTree,
-    blockchain: BlockChain,
+    blockchain: Blockchain,
     account_tree: AccountTree,
 }
 
@@ -109,7 +109,7 @@ impl State {
             nullifier_tree,
             // SAFETY: We assume the loaded MMR is valid and does not have more than u32::MAX
             // entries.
-            blockchain: BlockChain::from_mmr_unchecked(chain_mmr),
+            blockchain: Blockchain::from_mmr_unchecked(chain_mmr),
             account_tree,
         });
 
@@ -517,7 +517,7 @@ impl State {
             })
             .expect("DB should have returned the header of the batch reference block");
 
-        // The order doesn't matter for PartialBlockChain::new, so swap remove is fine.
+        // The order doesn't matter for PartialBlockchain::new, so swap remove is fine.
         let batch_reference_block_header = headers.swap_remove(header_index);
 
         // SAFETY: This should not error because:
@@ -525,7 +525,10 @@ impl State {
         // - so none of the block headers block numbers should exceed the chain length of the
         //   partial MMR,
         // - and we've added blocks to a BTreeSet, so there can be no duplicates.
-        let partial_block_chain = PartialBlockChain::new(partial_mmr, headers)
+        //
+        // We construct headers and partial MMR in concert, so they are consistent. This is why we
+        // can call the unchecked constructor.
+        let partial_block_chain = PartialBlockchain::new_unchecked(partial_mmr, headers)
             .expect("partial mmr and block headers should be consistent");
 
         Ok(BatchInputs {
@@ -659,7 +662,7 @@ impl State {
             })
             .expect("DB should have returned the header of the latest block header");
 
-        // The order doesn't matter for PartialBlockChain::new, so swap remove is fine.
+        // The order doesn't matter for PartialBlockchain::new, so swap remove is fine.
         let latest_block_header = headers.swap_remove(latest_block_header_index);
 
         // SAFETY: This should not error because:
@@ -667,7 +670,10 @@ impl State {
         // - so none of the block header's block numbers should exceed the chain length of the
         //   partial MMR,
         // - and we've added blocks to a BTreeSet, so there can be no duplicates.
-        let partial_block_chain = PartialBlockChain::new(partial_mmr, headers)
+        //
+        // We construct headers and partial MMR in concert, so they are consistent. This is why we
+        // can call the unchecked constructor.
+        let partial_block_chain = PartialBlockchain::new_unchecked(partial_mmr, headers)
             .expect("partial mmr and block headers should be consistent");
 
         Ok(BlockInputs::new(